+++ plans/.sprint12.md
@@ -130,10 +130,18 @@
 - TenantFilter bug → throws exception instead of sendError
 
 **Fixes Applied (3 commits):**
+
 - ✅ Extract async to `AsyncProvisioningExecutor` component
 - ✅ Add Guava RateLimiter (5 req/min per tenant, 429 response)
 - ✅ Add idempotency test + 11 TenantFilter unit tests
 - ✅ Fix TenantFilter `.orElseThrow()` → `.orElse(null)` with proper 400/403 errors
+
+**Frontend Refactoring (PR #69):**
+
+- ⚠️ provision-dialog: SRP violation (2 workflows, 649 LOC), 150 lines duplication, 11 getters
+- ✅ Split: ModuleProvisionDialogComponent + LanguageProvisionDialogComponent + ModuleCardComponent
+- ✅ Computed signals (0 getters), takeUntilDestroyed(), proper encapsulation
+- ✅ Result: -22% LOC, -100% duplication, -100% getters, +SRP compliance
 
 ## Later TODOs (post-sprint)
