+++ plans/SPRINT12_SUMMARY.md
@@ -268,4 +268,56 @@ Response: { "result": "SUCCESS", "data": [ { "code": "core", "name": "Core Modul
 
 **Final Status:** ✅ All review issues resolved, tests passing (11/11)
 
+## Post-Implementation Review (PR #69)
+
+**Frontend Refactoring - Provision Dialog Optimization:**
+
+- ⚠️ Single component with 2 workflows → SRP violation (649 LOC total)
+- ⚠️ 150 lines of template duplication (module list rendered 3x)
+- ⚠️ 11 getters instead of computed signals → performance issue
+- ⚠️ Parent-child coupling: TenantsListComponent mutates child data
+
+**Fixes Applied (PR #69 - Frontend Refactoring):**
+
+1. **Split into specialized components** (Component separation)
+   - `ModuleProvisionDialogComponent` (~300 LOC) - Module provisioning only
+   - `LanguageProvisionDialogComponent` (~205 LOC) - Language provisioning only
+   - `ModuleCardComponent` (40 LOC) - Reusable module card, eliminates 150 LOC duplication
+
+2. **Modern Angular patterns**
+   - ✅ 0 getters → All computed signals (memoized, performant)
+   - ✅ takeUntilDestroyed() for automatic cleanup (no memory leaks)
+   - ✅ OnPush change detection
+   - ✅ Signal-based inputs/outputs
+
+3. **Architecture improvements**
+   - ✅ Proper encapsulation: Polling logic moved to dialogs
+   - ✅ Single Responsibility Principle: Each component handles one workflow
+   - ✅ DRY principle: Module card extracted as reusable component
+   - ✅ Template separation: HTML in separate files
+
+4. **Code quality**
+   - ✅ No code comments (self-documenting code)
+   - ✅ Template reduction: 372 → 205 LOC (-45%)
+   - ✅ Improved maintainability and testability
+
+**Metrics:**
+
+| Metric | Before | After | Improvement |
+|--------|--------|-------|-------------|
+| Total LOC | 649 (1 component) | ~505 (3 components) | -22% |
+| Template duplication | 150 lines | 0 lines | -100% |
+| Getters | 11 | 0 | -100% |
+| Computed signals | 0 | 10+ | Modern pattern |
+| Components | 1 (mixed) | 3 (specialized) | SRP compliant |
+
+**Files Created:**
+
+- `storefront/src/app/shared/components/module-provision-dialog/*` (5 files)
+- `storefront/src/app/shared/components/language-provision-dialog/*` (4 files)
+
+**Files Deleted:**
+
+- `storefront/src/app/shared/components/provision-dialog/*` (old unified component)
+
