+++ storefront/src/app/shared/components/language-provision-dialog/language-provision-dialog.component.scss
@@ -2,5 +2,3 @@ mat-dialog-content {
     max-height: 70vh;
     overflow-y: auto;
 }
-
-
