+++ storefront/src/app/shared/components/module-provision-dialog/module-provision-dialog.component.scss
@@ -1,0 +1,4 @@
+mat-dialog-content {
+    max-height: 70vh;
+    overflow-y: auto;
+}
